* Change Athena::utility::fileSize to use stat64 instead of ftello * Fix output of ftello64 in win32_largefilewrapper

+++ include/win32_largefilewrapper.h
@@ -7,7 +7,7 @@
 extern "C" {
 #endif
 int fseeko64(FILE* fp, off64_t offset, int whence);
-int ftello64(FILE* fp);
+off64_t ftello64(FILE* fp);
 #ifdef __cplusplus
 }
 #endif
