+++ src/win32_largefilewrapper.c
@@ -7,7 +7,7 @@ int fseeko64(FILE* fp, off64_t offset, int whence)
     return _fseeki64(fp, offset, whence);
 }
 
-int ftello64(FILE* fp)
+off64_t ftello64(FILE* fp)
 {
     return _ftelli64(fp);
 }
